chore(clean-types): moves the types into types/wip whilst I work out the public API
