+++ runem/hook_manager.py
@@ -7,7 +7,7 @@ from runem.config_metadata import ConfigMetadata
 from runem.job import Job
 from runem.job_execute import job_execute
 from runem.log import log
-from runem.types import (
+from runem.types_wip.types import (
     FilePathListLookup,
     HookConfig,
     HookName,
