+++ runem/types_public_api.py
@@ -1,4 +1,4 @@
-from runem.types import FilePathList, JobName, OptionsWritable
+from runem.types_wip.types import FilePathList, JobName, OptionsWritable
 from runem.types_wip.types_jobs import JobKwargs, JobReturnData
 
 __all__ = ["FilePathList", "JobName", "JobReturnData", "OptionsWritable", "JobKwargs"]
