+++ scripts/test_hooks/json_validators.py
@@ -3,7 +3,7 @@ import pathlib
 from typing_extensions import Unpack
 
 from runem.run_command import run_command
-from runem.types import FilePathList
+from runem.types_wip.types import FilePathList
 from runem.types_wip.types_jobs import JobKwargs
