+++ tests/test_job_execute.py
@@ -13,7 +13,7 @@ from runem.config_metadata import ConfigMetadata
 from runem.informative_dict import InformativeDict
 from runem.job_execute import job_execute
 from runem.run_command import RecordSubJobTimeType
-from runem.types import FilePathListLookup, JobConfig, PhaseGroupedJobs
+from runem.types_wip.types import FilePathListLookup, JobConfig, PhaseGroupedJobs
 from tests.intentional_test_error import IntentionalTestError
