+++ tests/test_report.py
@@ -4,7 +4,7 @@ from contextlib import redirect_stdout
 from datetime import timedelta
 
 from runem.report import _print_reports_by_phase, report_on_run
-from runem.types import OrderedPhases
+from runem.types_wip.types import OrderedPhases
 from runem.types_wip.types_jobs import (
     JobReturn,
     JobRunMetadata,
